backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orion-range-core"
version = "0.1.0"
description = "Orion Range Core backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "proxmoxer>=2.0.0,<3.0.0",
  "requests>=2.31.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
addopts = "-q"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["migrations*"]
